refactor: 차트 컴포넌트 분리
